requirements-dev.txt
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1

black==23.7.0
flake8==6.0.0
isort==5.12.0
mypy==1.5.1

pre-commit==3.3.3
bandit==1.7.5
safety==2.3.4

jupyter==1.0.0
ipython==8.14.0
notebook==7.0.2

faker==19.6.2
factory-boy==3.3.0
responses==0.23.3

httpx==0.24.1
requests-mock==1.11.0
freezegun==1.2.2

coverage==7.3.0
pytest-html==3.2.0
pytest-xdist==3.3.1

python-dotenv==1.0.0
watchdog==3.0.0
python-multipart==0.0.6

uvicorn[standard]==0.23.2
gunicorn==21.2.0
fastapi==0.103.0

psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.1

sphinx==7.1.2
sphinx-rtd-theme==1.3.0
mkdocs==1.5.2
mkdocs-material==9.2.6

locust==2.16.1
memory-profiler==0.61.0
py-spy==0.3.14

docker==6.1.3
kubernetes==27.2.0
boto3==1.28.57

anthropic==0.3.11
openai==0.28.0
langchain==0.0.292
chromadb==0.4.10
tiktoken==0.4.0

numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

aiofiles==23.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

loguru==0.7.0
structlog==23.1.0
sentry-sdk==1.29.2

email-validator==2.0.0
pydantic[email]==2.3.0
sqlalchemy==2.0.20
alembic==1.11.3

python-dateutil==2.8.2
pytz==2023.3
croniter==1.4.1

click==8.1.7
typer==0.9.0
rich==13.5.2
tabulate==0.9.0